requirements.txt
# Minimal local RAG stack for PDF test cases
pymupdf==1.24.8          # text extraction (PyMuPDF)
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.8.2
